+++ web-server/util/index.js
@@ -1,0 +1,6 @@
+module.exports = {
+  R: require('ramda'),
+  error: require('./error'),
+  format: require('./formatters'),
+  fetchData: require('./fetch-data')
+}
